requirements.txt
flask
azure-identity
azure-mgmt-compute
azure-mgmt-resource